Move ViewModels into ViewModels namespace
