+++ visitz/Visitz/Views/Entity/EntityContactsView.xaml
@@ -2,8 +2,8 @@
 <views:ViewModelContentView 
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:entity="clr-namespace:Visitz.Views.Entity"
-    xmlns:models="clr-namespace:Visitz.Models"
+    xmlns:entity="clr-namespace:Visitz.ViewModels.Entity"
+    xmlns:models="clr-namespace:VisitzModel.Models;assembly=VisitzModel"
     xmlns:views="clr-namespace:Visitz.Views"
     x:Class="Visitz.Views.Entity.EntityContactsView"
     x:DataType="entity:EntityContactsViewModel">
