+++ visitz/Visitz/Views/Entity/EntityContainerView.xaml
@@ -4,7 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:views="clr-namespace:Visitz.Views"
     xmlns:tagview="clr-namespace:Visitz.Views.TagViews"
-    xmlns:entity="clr-namespace:Visitz.Views.Entity"
+    xmlns:entity="clr-namespace:Visitz.ViewModels.Entity"
     xmlns:tagBehaviors="clr-namespace:Visitz.Behaviors.Tags"
     x:Class="Visitz.Views.Entity.EntityContainerView"
     x:DataType="entity:EntityContainerViewModel">
