+++ visitz/Visitz/Views/Entity/EntityDetailsView.xaml
@@ -3,9 +3,9 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:views="clr-namespace:Visitz.Views"
-    xmlns:entity="clr-namespace:Visitz.Views.Entity"
+    xmlns:entity="clr-namespace:Visitz.ViewModels.Entity"
     xmlns:pages="clr-namespace:Visitz.Pages"
-    xmlns:models="clr-namespace:Visitz.Models"
+    xmlns:models="clr-namespace:VisitzModel.Models;assembly=VisitzModel"
     xmlns:fonticons="clr-namespace:Visitz.FontIcons"
     x:Class="Visitz.Views.Entity.EntityDetailsView"
     x:DataType="entity:EntityDetailsViewModel">
