+++ visitz/Visitz/Views/Navigation/HorizontalNavRailView.xaml
@@ -6,8 +6,9 @@
     xmlns:views="clr-namespace:Visitz.Views"
     xmlns:nav="clr-namespace:Visitz.Views.Navigation"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+    xmlns:viewModels="clr-namespace:Visitz.ViewModels"
     x:Class="Visitz.Views.Navigation.HorizontalNavRailView"
-    x:DataType="nav:NavRailViewModel"
+    x:DataType="viewModels:NavRailViewModel"
     BackgroundColor="{StaticResource BuildBarBackgroundColor}"
     MaximumHeightRequest="72">
