+++ visitz/Visitz/Views/Navigation/VerticalNavRailView.xaml
@@ -6,9 +6,10 @@
     xmlns:nav="clr-namespace:Visitz.Views.Navigation"
     xmlns:icons="clr-namespace:Visitz.FontIcons"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+    xmlns:viewModels="clr-namespace:Visitz.ViewModels"
     xmlns:views="clr-namespace:Visitz.Views"
     x:Class="Visitz.Views.Navigation.VerticalNavRailView"
-    x:DataType="nav:NavRailViewModel"
+    x:DataType="viewModels:NavRailViewModel"
     BackgroundColor="{StaticResource BuildBarBackgroundColor}"
     MaximumWidthRequest="90">
