+++ visitz/Visitz/VisitzConfig/VisitzScreens.cs
@@ -2,6 +2,7 @@
 using Visitz.Device;
 using Visitz.Pages;
 using Visitz.ViewModels;
+using Visitz.ViewModels.Entity;
 using Visitz.Views;
 using Visitz.Views.Caseload;
 using Visitz.Views.Debugging;
